Move internal files into +gitaly
Currently the database and staging directories for the
`PartitionManager` are housed in the root directory of a storage. This
may cause the `RepositoryCounter` to capture staged changes as
repositories during startup. In addition, historically Gitaly's internal
files for a storage have been kept in the `+gitaly` directory.

Move the `PartitionManager`'s directories into `+gitaly`.