storagemgr: Use correct snapshot LSN in traces
The snapshot LSN used for the transaction is retrieved from the
transaction manager after acquiring its mutex. However, this was being
read outside of the critical section when we set the trace span and
could result in a data race.

Set the trace span after the critical section, using the snapshot LSN
from the transaction object.