gittest: Create and expose `BackendSpecificRepoHash`
In the upcoming commits, we'll modify many tests which check the voting
hash of newly created repositories (when using transactions). Currently
they only run with the files backend, but when we add support for the
reftable backend, we need to ensure that they continue working.

Introduce a new test helper function to allow us to check the voting
hash of a repository, this helper function is a basic mock of how
repoutil.Create will be modified to work with the reftable backend.