testserver: Use GracefulStop in test cleanup
When we run Gitaly for tests, we start separate gRPC servers for
internal and external traffic. The internal server is used to handle
traffic from executing hooks. It's important to close the external
server before the internal one to prevent new requests from being
accepted. These requests may fail in-flight due to the internal server
being shut down.

Adjust the cleanup callback for the test server so we perform a graceful
shutdown for all servers spawned by the server factory. This simplifies
the code somewhat, and moves the responsibility of shutdown order to the
server factory.