Gitaly pod init container binary
The current containerized Gitaly pod does not have sufficient
permissions to create cgroups in the typical cgroups path for Kubernetes
pods. This makes it unable to manage cgroup (v2) controllers. Hence,
an init container is required to run first before the Gitaly pod to
change permissions on the host's cgroup path, so by the time Gitaly
pod runs, it will have the right permissions to write to cgroups.