test: Use absolute path for signing keys
A subsequent commit will begin to spawn Git commands from either the
repository it operates on, or a fixed scratch directory. This will cause
test failures due to the use of relative paths for various signing keys.

Add a test helper to compute the absolute path to the current test's
`testdata/` directory, and pass this to Git and Gitaly configuration
during test setup.