Don't sort test fixtures in place
TransactionManager's tests currently have a race in some of the test
cases that share a slice of expected object IDs in their assertions.
The list of expected object IDs is sorted prior to asserting for
equality to ensure there are no ordering problems. As some of the
tests share the same slice of expected object IDs, this leads to
racy access with the sorting writing into a shared slice. Always
return a new sorted slice instead to avoid this.