Move LSN type from storagemgr to storage
LSN type describes a location in a partition's write-ahead log.
Functionality such as backups require access to LSNs in order to
store the LSN of a given base backup or WAL entry. Move the type
from storagemgr package to the higher level storage package where
it can be imported from when needed.