gitaly-hooks: Introduce proc-receive hook
When `receive.procReceiveRefs` is configured, the proc-receive hook gets
invoked by git-receive-pack(1). This hook is responsible for updating
the relevant references and reporting the results back to receive-pack.

In Gitaly, the `ProcReceiveHook` RPC is intended to provide this hooks
functionality. Add the proc-receive hook to gitaly-hooks and set it up
to invoke the hook RPC.