git: Add proc-receive hook configuration
For gitaly-hooks to execute a specific hook, it must be requested in the
hooks payload. Add the proc-receive hook to the request hooks payload
when it is indicated the hooks should be enabled. Note that the command
configuration is also updated to include `receive.procReceiveRefs` so
git-receive-pack(1) will execute the proc-receive hook when enabled.

For now the hooks is always disabled, but this will change in a future
commit when the `ReceivePack` RPCs are configured to use the hook.