hook: Allow transaction to be commited already
For `ReceivePack` RPCs to use the proc-receive hook, the RPC will have to
commit its own transaction. This causes errors if the transaction is
committed again further downstream and could occur during the
post-receive hook. To avoid the error, allow the post-receive hooks to
fail committing if it is due the transaction already being committed.