Test pruning with an in-between reference update depending on object
Some transactions may consist of multiple reference transactions
performed by Git. When a transaction is being committed, we verify
all the objects it needs are present. As part of this, we need to
verify the intermediate objects a reference was pointed to during
a transaction exist as well, not just the final object.

We're currently missing a test asserting that the intermediate
objects are not allowed to be pruned by an pruning operation. Add
the missing test. The newly added test verifies the conflict checking
was done correctly so no changes are needed.