hook: Fix pkt-line header parsing
When the proc-receieve hook executes, the first pkt-line parsed is a
header containing the version and potentially a list of capabilities. If
the header contains capabilities a null character is appended to the
version string followed by a space separated list of capabilities. If no
capabilities are provided, no null character is appended.

The proc-receive hook pkt-line parser always expects a null character
after the version string even if there are no capabilities. This can
cause errors when parsing the pkt-line header.

Fix the pkt-line parser by not requiring a null character after the
version string.