git: Lenient pattern for MismatchingStateRegex
When `git update-ref` is executed by localrepo.UpdateRef, an optional
expected OID for the target reference can be provided. An error is
returned if the actual OID of the target doesn't match, for example if a
concurrent update occurred.

The error message emitted by Git in this scenario

    `fatal: cannot lock ref...`

doesn't match the regex pattern we have for MismatchingState (only one
colon is present in the error).

Modify the pattern with an optional non-capturing group so an error
without a second colon can still be matched.