git: Set attr.tree for git commands
In Git change 2386535511, we introduced a feature:
 "attr: read attributes from HEAD when bare repo".
This causes a performance degradation.
Current workaround is to get rid of the default behavior
in bare repos of reading from HEAD by setting attr.tree to empty tree.

This commit add a logic: when a command is in a white list, the
attr.tree is set to HEAD; otherwise the attr.tree is set to
an empty tree".

fixed: Bug fix