fetch: Add expected OID to prevent races
The `expected_old_oid` field has been introduced in many RPCs to help
prevent race conditions. This value indicates what the caller expects a
particular reference to point to _before_ it's updated, and is validated
by Git before the update occurs. If the value changes in the meantime,
an error is returned.

Add `expected_target_old_oid` to the `FetchSourceBranch` RPC. In this
case it indicates the OID that the target reference should point to
before a source branch is fetched into it. A FAILED_PRECONDITION gRPC
error is returned if the OID doesn't match.