receivepack: Add proc-receive hook helper
When the `ProcReceiveHook` RPC is invoked, a proc-receive handler is
transmitted to a waiting RPC handler to faciliate the hook execution.
The receive-pack RPC handlers need to use this handler to perform
reference updates and report which ones were accepted/rejected. In the
future, transactions will also need to be committed as part of the
proc-receive hook execution. To allow reuse between receive-pack RPCs,
`receivepack.RegisterProcReceiveHook()` is added to perform these shared
operations.