smarthttp: Use proc-receive hook in `PostReceivePack`
When invoked, the proc-receive hook becomes responsible for updating
references matching a configured pattern. Using this hook can allow for
more fine-grained control regarding when references are accepted or
rejected. Setup the `PostReceivePack` RPC to begin executing the
proc-receive hook when transactions are enabled. In a future commit, we
will take advantage of the proc-receive hook by committing the
transaction during hook execution.