Move key-value related types to their own package
We're soon going to be adding support for storing key-value data in
partitions. Doing so, we'll need to expose new key-value related
types. Currently the key-value store related functionality is in the
storagemgr package. This makes it difficult to give good names for the
types as the package's scope is so wide. Move all of the key-value
related functionality into their own package so we have a clean
namespace to use for the types. This way package names for example
differentiate storagemgr.Transaction and keyvalue.Transaction. This
also helps keeping the storagemgr package a bit smaller as it is
growing in size.