Use shorter key for partition prefix
The current partition key prefix is unnecessarily long. As it is
a database key, it's length will directly affect performance of
all read and write operations into the key-value store. Use a
shorter prefix for efficiency. We can later provide functionality
to convert the key names into human readable names if necessary.