config: Use an absolute path for SocketPath
Gitaly and Praefect can be configured to listen on a Unix socket or a
TCP port. This configuration is read from the respective `config.toml`
file.

Since the socket path is simply a file path, it's possible for Gitaly
and Praefect to be configured with a relative socket path. This happens
frequently in Rails specs, where the path `tmp/tests/gitaly/*.socket` is
used. This has not been problematic in the past because whenever the
socket path is referenced, the current working directory of the Gitaly
process has always been correct relative to that path.

With https://gitlab.com/gitlab-org/gitaly/-/issues/6065 however, we want
to restrict Git invocations to a specific working directory. Setting the
working directory of the command or using Git's `-C` flag both have the
side effect of causing test failures due to the relative socket path.

Fix this by computing the absolute path of the socket when the config
struct is created.