git: Ignore fsck warning for symlinks pointing to Git dir
The `symlinkPointsToGitDir` configuration was introduced in the recent
Git security release and by default now warns when the repository
contains a symlink that points inside the Git directory. If the
`transfer.fsckObjects` configuration is enabled, the warning instead
results in an error. Consequently, this breaks some existing
repositories. Disable this check by setting `symlinkPointsToGitDir` to
`ignore`.