git: Disable post-receive hook during transactions
In a future commit, the proc-receive hook will be enabled for
receive-pack RPCs while running in a transaction. During execution of
the proc-recieve hook the transaction will be committed. This requires
the subsequent post-receive hook to run in its own transaction. To
facilitate this, receive-pack RPC handlers need to invoke the
post-receive hook manually. Only enable the pre-receive and proc-receive
hooks when transactions are enabled.