Implement a key-value operation recorder
Transactions need to record the key-value operations being performed
so they can be conflict checked and write-ahead logged. Add a wrapper
type for ReadWriter that records all of the operations being performed.
Direct reads and writes are recorded. In addition, iterated prefixes
are recorded so they can be used for predicate locking.