Don't log errors when transaction is already committed
Transaction started in the middleware may be committed in the handler
if it invokes post-receive hook. This is because post-receive hook
needs to be invoked after transaction commit to ensure the data is
readable for any follow up requests.

We're currently logging an error if the transaction was already
committed when we're rolling it back. This is harmless but leads to
unnecessary error logs. Don't log the already committed errors.