Log more details about partitions
Gitaly is currently logging a repository's relative path instead
of a partition's ID when a partition fails. This is a remnant from
the past before partitions when each repository had its own
transaction manager. These days we should rather be logging the ID
of the partition. Include the partition ID as field in the logger.

Additionally, if the partition fails, include the state directory
of the partition as a field. This makes it easier to locate the
state of the partition having issues.