Don't retain write buffer when chunking responses
ListBlobs and ListAllBlobs are currently returning corrupted blobs
when the blob is large enough to require chunking. This is because
the chunkers are retaining the writers buffer. As the writer performs
the next write, it overwrites the exisiting data in the buffer. This
leads to the data in the previous chunks being overwritten with the
latest chunks data. Fix the issue by copying the data.

Changelog: fixed