Add a test case covering directory created by ref deletion
RecordReferenceUpdates currently has an issue where it doesn't
record directories created by reference deletions. As the created
directory is not logged, we may end up producing an incorrect WAL
entry that tries to create a child directory in it without first
creating the parent directory. Cover this case with a test.