gittest: Add function to skip test for a Git version
In the upcoming commit, we'll introduce a new way to update default
branches using 'symref-udpate' subcommand. This will be part of an
upcoming Git release and will rely on reference-transaction hook to run.

So existing tests will fail to run with this functionality, so we need
to skip them when running for this experimental Git version. Add a
function to allow us to do so.

Also introduce `IfSymrefUpdateSupported` which will help us pick the
write output for tests if the version supports symref-update.