Add the LogManagerAccessor interface
The `PartitionManager` will close a partition when it becomes idle. As a
result, it is possible for a `LogManager` to close before its
`LogConsumer` is able to acknowledge a completed transaction.

To ensure that the `LogManager` is active for acknowledgement, add a new
`LogManagerAccessor` interface that exposes the `CallLogManager`
function. This will start the `TransactionManager` for the requested
partition, if necessary, and then execute the provided function against
the manager.