Introduce LogManagerFactory type
We will shortly start storing the `PartitionManager` as struct member in
`LogEntryArchiver`, which creates a mutual dependency between these
types.

Introduce the `LogManagerFactory` to allow the `PartitionManager` to
create the `LogManager` as part of its own constructor. This also
provides a cleanup function to be executed by the `PartitionManager`
during its own shutdown process.