Add storage name to NotifyNewTransactions API
Currently the `NotifyNewTransactions` interface does not specify which
storage the partition being updated is associated with. This may lead
to collisions on Gitaly nodes where there are multiple active storages.

Add the storage name in `NotifyNewTransactions` and include it in
the archive path to ensure that one partition does not overwrite
another's backups.