Start using the LogManagerAccessor API
Convert the `LogEntryArchiver` over to the new `LogManagerAccessor` API
to perform callbacks to the `LogManager`. Previously, we attempted to
call back to the `LogManager` that most recently notified of a new
transaction, but this may close at any time, potentially causing lost
acknowledgements.

By using `LogManagerAccessor` we ensure that the `LogManager` is active
when an acknowledgement is sent.